SOLR-14613 reduce number of files by grouping property keys and property values in single files
