+++ solr/core/src/java/org/apache/solr/cluster/placement/impl/NonNodeMetricKeyImpl.java
@@ -15,20 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.solr.cluster.placement.impl.propertykey;
+package org.apache.solr.cluster.placement.impl;
 
+import com.google.common.base.Preconditions;
 import org.apache.solr.cluster.placement.Node;
-import org.apache.solr.cluster.placement.SyspropPropertyValue;
-import org.apache.solr.cluster.placement.impl.propertyvalue.SyspropPropertyValueImpl;
-import org.apache.solr.common.cloud.rule.ImplicitSnitch;
+import org.apache.solr.cluster.placement.PropertyKey;
+import org.apache.solr.cluster.placement.PropertyValueSource;
 
-public class SyspropKeyImpl extends AbstractNodePropertyKey {
-  public SyspropKeyImpl(Node node, String syspropName) {
-    super(node, ImplicitSnitch.SYSPROP + syspropName);
+public class NonNodeMetricKeyImpl implements PropertyKey {
+  private final String metricName;
+  private final PropertyValueSource metricSource;
+
+  public NonNodeMetricKeyImpl(PropertyValueSource metricSource, String metricName) {
+    Preconditions.checkState(!(metricSource instanceof Node), "Illegal argument type " + Node.class);
+    this.metricSource = metricSource;
+    this.metricName = metricName;
   }
 
   @Override
-  public SyspropPropertyValue getPropertyValueFromNodeValue(Object nodeValue) {
-    return new SyspropPropertyValueImpl(this, nodeValue);
+  public PropertyValueSource getPropertyValueSource() {
+    return metricSource;
   }
 }
