+++ solr/core/src/java/org/apache/solr/cluster/placement/impl/PropertyValueFetcherImpl.java
@@ -27,7 +27,6 @@ import org.apache.solr.cluster.placement.Node;
 import org.apache.solr.cluster.placement.PropertyKey;
 import org.apache.solr.cluster.placement.PropertyValue;
 import org.apache.solr.cluster.placement.PropertyValueFetcher;
-import org.apache.solr.cluster.placement.impl.propertykey.NodeSnitchPropertyKey;
 import org.apache.solr.common.SolrException;
 
 public class PropertyValueFetcherImpl implements PropertyValueFetcher {
@@ -40,7 +39,7 @@ public class PropertyValueFetcherImpl implements PropertyValueFetcher {
   @Override
   public Map<PropertyKey, PropertyValue> fetchProperties(Set<PropertyKey> props) {
     // Convert the plugin view of what's to fetch into the underlying Solr representation
-    Map<Node, Map<String, NodeSnitchPropertyKey>> nodeToTagsAndKeys = getTagsAndKeysByNode(props);
+    Map<Node, Map<String, AbstractNodePropertyKey>> nodeToTagsAndKeys = getTagsAndKeysByNode(props);
     // Fetch the data from the remote nodes
     Map<Node, Map<String, Object>> tagValues = getTagValuesForAllNodes(nodeToTagsAndKeys);
     // Convert fetched values to the plugin representation
@@ -54,25 +53,25 @@ public class PropertyValueFetcherImpl implements PropertyValueFetcher {
    * {@link org.apache.solr.client.solrj.cloud.NodeStateProvider#getNodeValues(String, Collection)} as well as which
    * {@link PropertyKey}'s they correspond to.
    */
-  private Map<Node, Map<String, NodeSnitchPropertyKey>> getTagsAndKeysByNode(Set<PropertyKey> props) {
+  private Map<Node, Map<String, AbstractNodePropertyKey>> getTagsAndKeysByNode(Set<PropertyKey> props) {
     // Map from a Node to a map from a tag name to the corresponding property key
-    Map<Node, Map<String, NodeSnitchPropertyKey>> nodeToTagsAndKeys = new HashMap<>();
+    Map<Node, Map<String, AbstractNodePropertyKey>> nodeToTagsAndKeys = new HashMap<>();
 
     for (PropertyKey propertyKey : props) {
-      if (propertyKey instanceof NodeSnitchPropertyKey) {
-        NodeSnitchPropertyKey nodeSnitchPropertyKey = (NodeSnitchPropertyKey) propertyKey;
-        Node node = nodeSnitchPropertyKey.getPropertyValueSource();
+      if (propertyKey instanceof AbstractNodePropertyKey) {
+        AbstractNodePropertyKey AbstractNodePropertyKey = (AbstractNodePropertyKey) propertyKey;
+        Node node = AbstractNodePropertyKey.getPropertyValueSource();
 
         // We might want to limit acceptable nodes here. We could verify if they're in the set of live nodes or similar
 
-        Map<String, NodeSnitchPropertyKey> tagsAndKeys = nodeToTagsAndKeys.get(node);
+        Map<String, AbstractNodePropertyKey> tagsAndKeys = nodeToTagsAndKeys.get(node);
         if (tagsAndKeys == null) {
           tagsAndKeys = new HashMap<>();
           nodeToTagsAndKeys.put(node, tagsAndKeys);
         }
 
-        String tag = nodeSnitchPropertyKey.getNodeSnitchTag();
-        tagsAndKeys.put(tag, nodeSnitchPropertyKey);
+        String tag = AbstractNodePropertyKey.getNodeSnitchTag();
+        tagsAndKeys.put(tag, AbstractNodePropertyKey);
       } else {
         // TODO support all keys...
         throw new SolrException(SolrException.ErrorCode.SERVER_ERROR, "Unsupported property key " + propertyKey.getClass().getName());
@@ -86,13 +85,13 @@ public class PropertyValueFetcherImpl implements PropertyValueFetcher {
    * Gets the tag values from remote nodes for all nodes. Returns {@link Object}'s as returned by
    * {@link org.apache.solr.client.solrj.cloud.NodeStateProvider#getNodeValues(String, Collection)}.
    */
-  private Map<Node, Map<String, Object>> getTagValuesForAllNodes(Map<Node, Map<String, NodeSnitchPropertyKey>> nodeToTagsAndKeys) {
+  private Map<Node, Map<String, Object>> getTagValuesForAllNodes(Map<Node, Map<String, AbstractNodePropertyKey>> nodeToTagsAndKeys) {
 
     Map<Node, Map<String, Object>> nodeToTagValues = new HashMap<>(nodeToTagsAndKeys.size());
 
     // This implementation does a sequential fetch from all nodes. This will most likely have to be converted to
     // parallel fetching from all nodes concurrently.
-    for (Map.Entry<Node, Map<String, NodeSnitchPropertyKey>> e : nodeToTagsAndKeys.entrySet()) {
+    for (Map.Entry<Node, Map<String, AbstractNodePropertyKey>> e : nodeToTagsAndKeys.entrySet()) {
       Node node = e.getKey();
       Map<String, Object> tagValues = cloudManager.getNodeStateProvider().getNodeValues(node.getNodeName(), e.getValue().keySet());
       if (tagValues != null) {
@@ -107,7 +106,7 @@ public class PropertyValueFetcherImpl implements PropertyValueFetcher {
    * Do the conversion of returned tag values (that are {@link Object}'s) into the strongly typed corresponding {@link PropertyValue}'s.
    * Returns a map with they keys being the original {@link PropertyKey}, for those properties that do have returned node values.
    */
-  private Map<PropertyKey, PropertyValue> convertToProperyValues(Map<Node, Map<String, NodeSnitchPropertyKey>> nodeToTagsAndKeys, Map<Node, Map<String, Object>> nodeToTagValues) {
+  private Map<PropertyKey, PropertyValue> convertToProperyValues(Map<Node, Map<String, AbstractNodePropertyKey>> nodeToTagsAndKeys, Map<Node, Map<String, Object>> nodeToTagValues) {
     Map<PropertyKey, PropertyValue> keysToValues = new HashMap<>();
 
     for (Map.Entry<Node, Map<String, Object>> e : nodeToTagValues.entrySet()) {
@@ -115,16 +114,16 @@ public class PropertyValueFetcherImpl implements PropertyValueFetcher {
       // PropertyKey's (tagToPropertyKey).
       Node node = e.getKey();
       Map<String, Object> tagToNodeValue = e.getValue();
-      Map<String, NodeSnitchPropertyKey> tagToPropertyKey = nodeToTagsAndKeys.get(node);
+      Map<String, AbstractNodePropertyKey> tagToPropertyKey = nodeToTagsAndKeys.get(node);
 
       // Iterating over the requested property keys we find the corresponding tag node value (if it exists) and use the
       // property key to cast that value into the correct type and return the corresponding PropertyValue.
-      for (Map.Entry<String, NodeSnitchPropertyKey> tp : tagToPropertyKey.entrySet()) {
-        NodeSnitchPropertyKey nodeSnitchPropertyKey = tp.getValue();
-        Object nodeValue = tagToNodeValue.get(nodeSnitchPropertyKey.getNodeSnitchTag());
+      for (Map.Entry<String, AbstractNodePropertyKey> tp : tagToPropertyKey.entrySet()) {
+        AbstractNodePropertyKey AbstractNodePropertyKey = tp.getValue();
+        Object nodeValue = tagToNodeValue.get(AbstractNodePropertyKey.getNodeSnitchTag());
         if (nodeValue != null) {
-          PropertyValue pv = nodeSnitchPropertyKey.getPropertyValueFromNodeValue(nodeValue);
-          keysToValues.put(nodeSnitchPropertyKey, pv);
+          PropertyValue pv = AbstractNodePropertyKey.getPropertyValueFromNodeValue(nodeValue);
+          keysToValues.put(AbstractNodePropertyKey, pv);
         }
       }
     }
