--- solr/core/src/java/org/apache/solr/cluster/placement/impl/propertyvalue/SyspropPropertyValueImpl.java
@@ -1,37 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.solr.cluster.placement.impl.propertyvalue;
-
-import org.apache.solr.cluster.placement.PropertyKey;
-import org.apache.solr.cluster.placement.SyspropPropertyValue;
-import org.apache.solr.cluster.placement.SystemLoadPropertyValue;
-import org.apache.solr.cluster.placement.TotalDiskPropertyValue;
-
-public class SyspropPropertyValueImpl extends AbstractPropertyValue implements SyspropPropertyValue {
-  private final String systemProperty;
-
-  public SyspropPropertyValueImpl(PropertyKey key, Object nodeValue) {
-    super(key);
-    this.systemProperty = (String) nodeValue;
-  }
-
-  @Override
-  public String getSystemPropertyValue() {
-    return systemProperty;
-  }
-}
