--- solr/core/src/java/org/apache/solr/cluster/placement/impl/propertyvalue/TotalDiskPropertyValueImpl.java
@@ -1,36 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.solr.cluster.placement.impl.propertyvalue;
-
-import org.apache.solr.cluster.placement.FreeDiskPropertyValue;
-import org.apache.solr.cluster.placement.PropertyKey;
-import org.apache.solr.cluster.placement.TotalDiskPropertyValue;
-
-public class TotalDiskPropertyValueImpl extends AbstractPropertyValue implements TotalDiskPropertyValue {
-  private final long totalDiskGB;
-
-  public TotalDiskPropertyValueImpl(PropertyKey key, Object nodeValue) {
-    super(key);
-    this.totalDiskGB = ((Number) nodeValue).longValue();
-  }
-
-  @Override
-  public long getTotalSizeGB() {
-    return totalDiskGB;
-  }
-}
